pyproject.toml
[project]
name = "mods"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "discord-py>=2.6.4",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
]

[dependency-groups]
dev = [
    "alembic>=1.17.2",
]

